rework(vendors)
Started reworking process of old and malfunctioning vendor plugin.
Added new types of vendor access: ID-based and password.
ID access scans ID card (if equipped) of the one attempting to interact with vendor. If the card has required access (which is set by admins via context menu), person is allowed to interact with vendor. If not, other checks are performed. Password system was also added as an alternative access method. Password is set by admins via context menu. In this menu, there is user-accessible button which allows them to enter password. Entering correct password will allow user to interact with vendor regardless of any other restrictions for the remainder of the session.